(BIDS-2368) return withdrawals count func to db package
